+++ src/components/value/index.tsx
@@ -4,7 +4,7 @@ import { parseOperandTextExpr, parseTextExpr } from '../../utils/parseTextExpr';
 import { SDMXParser } from 'sdmx-json-parser';
 import { parseDataExpr } from "../../utils/parseDataExpr";
 
-const Value = ({ config }: { config: any }) => {
+const Value = ({ config, language }: { config: any, language: string }) => {
 
     const [titleObject, setTitleObject] = useState<any>({ text: "Loading..." })
     const [subTitleObject, setSubTitleObject] = useState<any>({ text: "Loading..." })
@@ -12,7 +12,7 @@ const Value = ({ config }: { config: any }) => {
 
     const sdmxParser = new SDMXParser();
 
-    const formatValue = (valueStr: any, config: any, data: any, attributes: any) => {
+    const formatValue = (valueStr: any, config: any, data: any, attributes: any, language: string) => {
         if (config['Decimals']) {
             const decimalNumber = Number(parseOperandTextExpr(config['Decimals'], data[0], attributes));
             valueStr = Number(valueStr).toFixed(decimalNumber);
@@ -24,11 +24,11 @@ const Value = ({ config }: { config: any }) => {
                 valueStr = config['Unit'] + valueStr;
             }
         }
-        return valueStr.toLocaleString();
+        return valueStr.toLocaleString(language);
     }
 
     useEffect(() => {
-        const dataObjs = parseDataExpr(config.DATA);
+        const dataObjs = parseDataExpr(config.data);
         if (dataObjs.length > 1) {
             throw new Error('Multiple data expressions are not supported for Value component');
         }
@@ -38,16 +38,15 @@ const Value = ({ config }: { config: any }) => {
         sdmxParser.getDatasets(dataFlowUrl, {
             headers: new Headers({
                 Accept: "application/vnd.sdmx.data+json;version=2.0.0",
+                "Accept-Language": language
             })
         }).then(() => {
             const data = sdmxParser.getData();
             const dimensions = sdmxParser.getDimensions();
             const attributes = sdmxParser.getAttributes();
 
-            const titleObj = parseTextExpr(config.Title, dimensions);
-            const subTitleObj = parseTextExpr(config.Subtitle, dimensions);
-            setTitleObject(titleObj);
-            setSubTitleObject(subTitleObj);
+            setTitleObject(config.title)
+            setSubTitleObject(config.subtitle)
 
             let valueStr = data[0].value;
             // apply operation to data
@@ -56,7 +55,7 @@ const Value = ({ config }: { config: any }) => {
                     // if operand starts with { then it is an attribute
                     const operandValue = parseOperandTextExpr(dataObj.operand, data[0], attributes);
                     valueStr = eval(`${valueStr} ${dataObj.operator} ${operandValue}`);
-                    setValueStr(formatValue(valueStr, config, data, attributes));
+                    setValueStr(formatValue(valueStr, config, data, attributes, language));
                 } else {
                     // we presume it is a dataflow url
                     const parserOperand = new SDMXParser();
@@ -68,20 +67,20 @@ const Value = ({ config }: { config: any }) => {
                         const dataOperand = parserOperand.getData();
                         const dataOperandValue = dataOperand[0].value;
                         valueStr = eval(`${valueStr} ${dataObj.operator} ${dataOperandValue}`);
-                        setValueStr(formatValue(valueStr, config, data, attributes));
+                        setValueStr(formatValue(valueStr, config, data, attributes, language));
                     });
                 }
 
             } else {
-                setValueStr(formatValue(valueStr, config, data, attributes));
+                setValueStr(formatValue(valueStr, config, data, attributes, language));
             }
         });
-    }, []);
+    }, [language]);
 
     return (
         <div className={`pt-3 pb-2 px-2 px-xl-3 bg-white h-100 d-flex flex-column min-cell-height ${config.frameYN && config.frameYN.toLowerCase() == 'yes' ? "border" : ""}`}>
-            <h2 className={`${titleObject.bootstrapcss && titleObject.bootstrapcss.join(' ')}`} style={titleObject.inlinecss}>{titleObject.text}</h2>
-            {subTitleObject['text'] && (<h4 className={`${subTitleObject.bootstrapcss && subTitleObject.bootstrapcss.join(' ')}`} style={subTitleObject.inlinecss}>{subTitleObject.text}</h4>)}
+            <h2 className={`${titleObject.bootstrapcss && titleObject.bootstrapcss.join(' ')}`} style={titleObject.inlinecss}>{titleObject.text[language]}</h2>
+            {subTitleObject['text'] && (<h4 className={`${subTitleObject.bootstrapcss && subTitleObject.bootstrapcss.join(' ')}`} style={subTitleObject.inlinecss}>{subTitleObject.text[language]}</h4>)}
             <div className="display-2 flex-grow-1 d-flex align-items-center justify-content-center">
                 <p>{valueStr}</p>
             </div>
